requirements.txt
absl-py==0.12.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
appnope==0.1.2
backcall==0.2.0
black==21.4b2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
decorator==5.0.9
detectron2==0.4
earthengine-api==0.1.266
future==0.18.2
fvcore==0.1.5.post20210518
google-api-core==1.28.0
google-api-python-client==1.12.8
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
httplib2shim==0.0.3
hydra-core==1.1.0rc1
idna==2.10
importlib-metadata==4.0.1
importlib-resources==5.1.4
iopath==0.1.8
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
numpy==1.20.3
oauthlib==3.1.0
omegaconf==2.1.0rc1
opencv-python==4.5.2.52
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.3.0
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.4.8
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.1
yacs==0.1.8
zipp==3.4.1
